refactor (listRoutes) separate new routes from those under toggle + make package for lists e2e tests
