+++ webapi/src/main/scala/org/knora/webapi/routing/admin/ListsRouteADM.scala
@@ -32,10 +32,12 @@ import org.knora.webapi.routing.{KnoraRoute, KnoraRouteData}
  */
 class ListsRouteADM(routeData: KnoraRouteData) extends KnoraRoute(routeData) {
     private val featureFactory: ListsRouteADMFeatureFactory = new ListsRouteADMFeatureFactory(routeData)
-    private val deleteRoute: DeleteListItemsRouteADM = new DeleteListItemsRouteADM(routeData)
+    private val deleteNodeRoute: DeleteListItemsRouteADM = new DeleteListItemsRouteADM(routeData)
+    private val updateNodeRoute: UpdateListItemsRouteADM = new UpdateListItemsRouteADM(routeData)
 
     override def makeRoute(featureFactoryConfig: FeatureFactoryConfig): Route = {
-        featureFactory.makeRoute(featureFactoryConfig)~
-        deleteRoute.makeRoute(featureFactoryConfig)
+        featureFactory.makeRoute(featureFactoryConfig) ~
+        deleteNodeRoute.makeRoute(featureFactoryConfig) ~
+        updateNodeRoute.makeRoute(featureFactoryConfig)
     }
 }
