+++ webapi/src/main/scala/org/knora/webapi/routing/admin/lists/NewListsRouteADMFeature.scala
@@ -51,13 +51,10 @@ class NewListsRouteADMFeature(routeData: KnoraRouteData) extends KnoraRoute(rout
 
     def makeRoute(featureFactoryConfig: FeatureFactoryConfig): Route =
         getLists(featureFactoryConfig) ~
-            createListItem(featureFactoryConfig) ~
-            getListItem(featureFactoryConfig) ~
-            updateListItem(featureFactoryConfig) ~
-            getNodeInfo(featureFactoryConfig) ~
-            updateNodeName(featureFactoryConfig) ~
-            updateNodeLabels(featureFactoryConfig) ~
-            updateNodeComments(featureFactoryConfig)
+        createListItem(featureFactoryConfig) ~
+        getListItem(featureFactoryConfig) ~
+        updateListItem(featureFactoryConfig) ~
+        getNodeInfo(featureFactoryConfig)
 
     /* return all lists optionally filtered by project */
     @Path("/{IRI}")
@@ -253,132 +250,6 @@ class NewListsRouteADMFeature(routeData: KnoraRouteData) extends KnoraRoute(rout
                     responderManager = responderManager,
                     log = log
                 )
-        }
-    }
-    /**
-     * update node name
-     */
-    @Path("/{IRI}/name")
-    @ApiOperation(value = "Update Node Name", nickname = "newPutNodeName", httpMethod = "PUT", response = classOf[NodeInfoGetResponseADM])
-    @ApiImplicitParams(Array(
-        new ApiImplicitParam(name = "body", value = "\"node name\" to update", required = true,
-            dataTypeClass = classOf[ChangeNodeNameApiRequestADM], paramType = "body"),
-        new ApiImplicitParam(name = "X-Knora-Feature-Toggles", value = "new-list-admin-routes:1 = on/off", required = true,
-            dataType = "string", paramType = "header")
-    ))
-    @ApiResponses(Array(
-        new ApiResponse(code = 500, message = "Internal server error")
-    ))
-    private def updateNodeName(featureFactoryConfig: FeatureFactoryConfig): Route = path(ListsBasePath / Segment / "name") { iri =>
-        put {
-            /* update name of an existing list node (either root or child) */
-            entity(as[ChangeNodeNameApiRequestADM]) { apiRequest =>
-                requestContext =>
-                    val nodeIri = stringFormatter.validateAndEscapeIri(iri, throw BadRequestException(s"Invalid param node IRI: $iri"))
-
-                    val requestMessage: Future[NodeNameChangeRequestADM] = for {
-                        requestingUser <- getUserADM(requestContext, featureFactoryConfig)
-                    } yield NodeNameChangeRequestADM(
-                        nodeIri = nodeIri,
-                        changeNodeNameRequest = apiRequest,
-                        featureFactoryConfig = featureFactoryConfig,
-                        requestingUser = requestingUser,
-                        apiRequestID = UUID.randomUUID()
-                    )
-
-                    RouteUtilADM.runJsonRoute(
-                        requestMessageF = requestMessage,
-                        requestContext = requestContext,
-                        featureFactoryConfig = featureFactoryConfig,
-                        settings = settings,
-                        responderManager = responderManager,
-                        log = log
-                    )
-            }
-        }
-    }
-    /**
-     * update node labels
-     */
-    @Path("/{IRI}/labels")
-    @ApiOperation(value = "Update Node Labels", nickname = "newPutNodeLabels", httpMethod = "PUT", response = classOf[NodeInfoGetResponseADM])
-    @ApiImplicitParams(Array(
-        new ApiImplicitParam(name = "body", value = "\"node labels\" to update", required = true,
-            dataTypeClass = classOf[ChangeNodeLabelsApiRequestADM], paramType = "body"),
-        new ApiImplicitParam(name = "X-Knora-Feature-Toggles", value = "new-list-admin-routes:1 = on/off", required = true,
-            dataType = "string", paramType = "header")
-    ))
-    @ApiResponses(Array(
-        new ApiResponse(code = 500, message = "Internal server error")
-    ))
-    private def updateNodeLabels(featureFactoryConfig: FeatureFactoryConfig): Route = path(ListsBasePath / Segment / "labels") { iri =>
-        put {
-            /* update labels of an existing list node (either root or child) */
-            entity(as[ChangeNodeLabelsApiRequestADM]) { apiRequest =>
-                requestContext =>
-                    val nodeIri = stringFormatter.validateAndEscapeIri(iri, throw BadRequestException(s"Invalid param node IRI: $iri"))
-
-                    val requestMessage: Future[NodeLabelsChangeRequestADM] = for {
-                        requestingUser <- getUserADM(requestContext, featureFactoryConfig)
-                    } yield NodeLabelsChangeRequestADM(
-                        nodeIri = nodeIri,
-                        changeNodeLabelsRequest = apiRequest,
-                        featureFactoryConfig = featureFactoryConfig,
-                        requestingUser = requestingUser,
-                        apiRequestID = UUID.randomUUID()
-                    )
-
-                    RouteUtilADM.runJsonRoute(
-                        requestMessageF = requestMessage,
-                        requestContext = requestContext,
-                        featureFactoryConfig = featureFactoryConfig,
-                        settings = settings,
-                        responderManager = responderManager,
-                        log = log
-                    )
-            }
-        }
-    }
-    /**
-     * update node comments
-     */
-    @Path("/{IRI}/comments")
-    @ApiOperation(value = "Update Node Comments", nickname = "newPutNodeComments", httpMethod = "PUT", response = classOf[NodeInfoGetResponseADM])
-    @ApiImplicitParams(Array(
-        new ApiImplicitParam(name = "body", value = "\"node comments\" to update", required = true,
-            dataTypeClass = classOf[ChangeNodeCommentsApiRequestADM], paramType = "body"),
-        new ApiImplicitParam(name = "X-Knora-Feature-Toggles", value = "new-list-admin-routes:1 = on/off", required = true,
-            dataType = "string", paramType = "header")
-    ))
-    @ApiResponses(Array(
-        new ApiResponse(code = 500, message = "Internal server error")
-    ))
-    private def updateNodeComments(featureFactoryConfig: FeatureFactoryConfig): Route = path(ListsBasePath / Segment / "comments") { iri =>
-        put {
-            /* update labels of an existing list node (either root or child) */
-            entity(as[ChangeNodeCommentsApiRequestADM]) { apiRequest =>
-                requestContext =>
-                    val nodeIri = stringFormatter.validateAndEscapeIri(iri, throw BadRequestException(s"Invalid param node IRI: $iri"))
-
-                    val requestMessage: Future[NodeCommentsChangeRequestADM] = for {
-                        requestingUser <- getUserADM(requestContext, featureFactoryConfig)
-                    } yield NodeCommentsChangeRequestADM(
-                        nodeIri = nodeIri,
-                        changeNodeCommentsRequest = apiRequest,
-                        featureFactoryConfig = featureFactoryConfig,
-                        requestingUser = requestingUser,
-                        apiRequestID = UUID.randomUUID()
-                    )
-
-                    RouteUtilADM.runJsonRoute(
-                        requestMessageF = requestMessage,
-                        requestContext = requestContext,
-                        featureFactoryConfig = featureFactoryConfig,
-                        settings = settings,
-                        responderManager = responderManager,
-                        log = log
-                    )
-            }
         }
     }
 }
