+++ webapi/src/test/scala/org/knora/webapi/e2e/admin/lists/BUILD.bazel
@@ -4,10 +4,10 @@ load("@io_bazel_rules_scala//scala:scala.bzl", "scala_test")
 load("//third_party:dependencies.bzl", "ALL_WEBAPI_MAIN_DEPENDENCIES", "BASE_TEST_DEPENDENCIES", "BASE_TEST_DEPENDENCIES_WITH_JSON")
 
 scala_test(
-    name = "GroupsADME2ESpec",
+    name = "OldListsRouteADMFeatureE2ESpec",
     size = "small",  # 60s
     srcs = [
-        "GroupsADME2ESpec.scala",
+        "OldListsRouteADMFeatureE2ESpec.scala",
         "//webapi/src/test/scala/org/knora/webapi/messages:SessionMessagesV1",
     ],
     data = [
@@ -23,48 +23,10 @@ scala_test(
 )
 
 scala_test(
-    name = "PermissionsADME2ESpec",
+    name = "NewListsRoutesADMFeatureE2ESpec",
     size = "small",  # 60s
     srcs = [
-        "PermissionsADME2ESpec.scala",
-    ],
-    data = [
-        "//knora-ontologies",
-        "//test_data",
-    ],
-    jvm_flags = ["-Dconfig.resource=fuseki.conf"],
-    # unused_dependency_checker_mode = "warn",
-    deps = ALL_WEBAPI_MAIN_DEPENDENCIES + [
-        "//webapi:main_library",
-        "//webapi:test_library",
-    ] + BASE_TEST_DEPENDENCIES_WITH_JSON,
-)
-
-scala_test(
-    name = "ProjectsADME2ESpec",
-    size = "small",  # 60s
-    srcs = [
-        "ProjectsADME2ESpec.scala",
-        "//webapi/src/test/scala/org/knora/webapi/messages:SessionMessagesV1",
-    ],
-    data = [
-        "//knora-ontologies",
-        "//test_data",
-    ],
-    jvm_flags = ["-Dconfig.resource=fuseki.conf"],
-    # unused_dependency_checker_mode = "warn",
-    deps = ALL_WEBAPI_MAIN_DEPENDENCIES + [
-        "//webapi:main_library",
-        "//webapi:test_library",
-        "@maven//:org_eclipse_rdf4j_rdf4j_client",
-    ] + BASE_TEST_DEPENDENCIES_WITH_JSON,
-)
-
-scala_test(
-    name = "SipiADME2ESpec",
-    size = "small",  # 60s
-    srcs = [
-        "SipiADME2ESpec.scala",
+        "NewListsRoutesADMFeatureE2ESpec.scala",
         "//webapi/src/test/scala/org/knora/webapi/messages:SessionMessagesV1",
     ],
     data = [
@@ -80,10 +42,11 @@ scala_test(
 )
 
 scala_test(
-    name = "StoreADME2ESpec",
+    name = "DeleteListItemsRouteADME2ESpec",
     size = "small",  # 60s
     srcs = [
-        "StoreADME2ESpec.scala",
+        "DeleteListItemsRouteADME2ESpec.scala",
+        "//webapi/src/test/scala/org/knora/webapi/messages:SessionMessagesV1",
     ],
     data = [
         "//knora-ontologies",
@@ -98,10 +61,10 @@ scala_test(
 )
 
 scala_test(
-    name = "UsersADME2ESpec",
+    name = "UpdateListItemsRouteADME2ESpec",
     size = "small",  # 60s
     srcs = [
-        "UsersADME2ESpec.scala",
+        "UpdateListItemsRouteADME2ESpec.scala",
         "//webapi/src/test/scala/org/knora/webapi/messages:SessionMessagesV1",
     ],
     data = [
